# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  dto.BorrowingInfoResponse:
    properties:
      currentInterestRate:
        type: string
      interestAccrued:
        type: string
      totalBorrowed:
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  dto.TransactionListResponse:
    properties:
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
      totalPage:
        type: integer
      transactions:
        items:
          $ref: '#/definitions/dto.TransactionResponse'
        type: array
    type: object
  dto.TransactionRequest:
    properties:
      amount:
        type: string
    required:
    - amount
    type: object
  dto.TransactionResponse:
    properties:
      amount:
        type: string
      blockNumber:
        type: integer
      createdAt:
        type: string
      gasPrice:
        type: string
      gasUsed:
        type: integer
      hash:
        type: string
      id:
        type: integer
      status:
        $ref: '#/definitions/dto.TransactionStatus'
      tokenAddress:
        type: string
      type:
        $ref: '#/definitions/dto.TransactionType'
      updatedAt:
        type: string
      userId:
        type: integer
    type: object
  dto.TransactionStatus:
    enum:
    - pending
    - confirmed
    - failed
    type: string
    x-enum-varnames:
    - TransactionStatusPending
    - TransactionStatusConfirmed
    - TransactionStatusFailed
  dto.TransactionType:
    enum:
    - deposit
    - withdraw
    - borrow
    - repay
    - liquidate
    type: string
    x-enum-varnames:
    - TransactionTypeDeposit
    - TransactionTypeWithdraw
    - TransactionTypeBorrow
    - TransactionTypeRepay
    - TransactionTypeLiquidate
...
host: localhost:8080
info:
  contact: {}
  description: API for decentralized lending and borrowing platform
  title: Lending & Borrowing Platform API
  version: "1.0"
paths:
  /borrowing/balance:
    get:
      consumes:
      - application/json
      description: Get the total amount borrowed by the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get borrowed amount
      tags:
      - borrowing
  /borrowing/borrow:
    post:
      consumes:
      - application/json
      description: Borrow tokens against deposited collateral
      parameters:
      - description: Borrow amount
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.TransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Borrow tokens against collateral
      tags:
      - borrowing
  /borrowing/info:
    get:
      consumes:
      - application/json
      description: Get detailed information about user's borrowing including interest
        accrued
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BorrowingInfoResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get borrowing information
      tags:
      - borrowing
  /borrowing/repay:
    post:
      consumes:
      - application/json
      description: Repay tokens previously borrowed from the protocol
      parameters:
      - description: Repay amount
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.TransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Repay borrowed tokens
      tags:
      - borrowing
  /borrowing/stats:
    get:
      consumes:
      - application/json
      description: Get global statistics about borrowing on the platform
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get borrowing statistics
      tags:
      - borrowing
  /borrowing/transactions:
    get:
      consumes:
      - application/json
      description: Get paginated history of user's borrowing transactions
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 10, max: 100)'
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TransactionListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get borrowing transaction history
      tags:
      - borrowing
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"